fixed the PR to follow atomic design, fixed other bugs so the app build
