+++ src/app/register/page.tsx
@@ -1,0 +1,5 @@
+import PatientRegistrationPage from "@/components/pages/patient-registration/patient-registration-page"
+
+export default function Register() {
+  return <PatientRegistrationPage />
+}
