+++ src/components/molecules/form-field.tsx
@@ -2,8 +2,8 @@
 
 import type React from "react"
 import type { ReactNode } from "react"
-import { Label } from "@/components/ui/label"
-import { Input } from "@/components/ui/input"
+import { Label } from "@/components/atoms/ui/label"
+import { Input } from "@/components/atoms/ui/input"
 
 interface FormFieldProps {
   label: string
