--- src/components/pages/page.tsx
@@ -1,5 +1,0 @@
-import PatientRegistrationPage from "@/components/templates/patient-registration-page"
-
-export default function Home() {
-  return <PatientRegistrationPage />
-}
